requirements.txt
openai
gTTS
moviepy==1.0.3
python-dotenv
